+++ CHANGELOG.md
@@ -8,6 +8,19 @@
     * `valueFragmentType`
     * `valueFragmentSeries`
 
+* Removed deprecated commands
+    `Remove-AuditRecordCollection`
+
+* Updated Pester module to v5
+
+* Fixed `New-Microservice` cmdlet
+
+* Updated `go-c8y` dependency to 0.8.0
+
+* Enabled ci test runner
+
+* Added parallel powershell tests execution to reduce test execution times
+
 ## Released
 
 ### v1.2.0